+++ src/config.rs
@@ -11,7 +11,6 @@ use std::{
 	fmt::{self, Debug, Display, Formatter},
 	fs, mem,
 	path::{Path, PathBuf},
-	str::FromStr,
 	sync::{RwLock, RwLockReadGuard, RwLockWriteGuard},
 };
 use toml;
@@ -28,14 +27,6 @@ pub enum ConfigKind {
 	Default,
 	Global(PathBuf),
 	Workspace(PathBuf),
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Deserialize, Serialize)]
-pub enum LineEnding {
-	#[default]
-	LF,
-	CR,
-	CRLF,
 }
 
 #[optfield(OptConfig, merge_fn, attrs = (derive(Deserialize)))]
@@ -98,8 +89,8 @@ pub struct Config {
 
 	/// Use .lua file extension instead of .luau when writing scripts
 	pub lua_extension: bool,
-	/// Line ending to use when writing Argon-generated files (LF, CRLF, CR)
-	pub line_ending: LineEnding,
+	/// Ignore line endings when reading files to avoid script diffs
+	pub ignore_line_endings: bool,
 	/// Package manager to use when running roblox-ts scripts (npm, bun, etc.)
 	pub package_manager: String,
 	/// Share anonymous Argon usage statistics with the community
@@ -144,7 +135,7 @@ impl Default for Config {
 			max_unsynced_changes: 10,
 
 			lua_extension: false,
-			line_ending: LineEnding::LF,
+			ignore_line_endings: true,
 			package_manager: String::from("npm"),
 			share_stats: true,
 
@@ -315,25 +306,6 @@ impl Display for ConfigKind {
 				Self::Workspace(_) => "Workspace",
 			}
 		)
-	}
-}
-
-impl Display for LineEnding {
-	fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-		write!(f, "{:?}", self)
-	}
-}
-
-impl FromStr for LineEnding {
-	type Err = String;
-
-	fn from_str(s: &str) -> Result<Self, Self::Err> {
-		match s.to_uppercase().as_str() {
-			"LF" => Ok(Self::LF),
-			"CR" => Ok(Self::CR),
-			"CRLF" => Ok(Self::CRLF),
-			_ => Err(format!("Invalid line ending: {}", s)),
-		}
 	}
 }